chore: Renamed files and removed the console log.
