+++ src/Pages/AppraisalActivity/datePicker.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import DateView from 'react-datepicker'
 import { Field, ErrorMessage } from 'formik'
-import TextError from './textError'
+import TextError from './TextError'
 import 'react-datepicker/dist/react-datepicker.css'
 
 function DatePicker (props) {
