+++ src/Pages/AppraisalActivity/input.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Field, ErrorMessage } from 'formik'
-import TextError from './textError'
+import TextError from './TextError'
 
 function Input (props) {
   const { label, name, ...rest } = props
