+++ src/Pages/AppraisalActivity/textArea.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Field, ErrorMessage } from 'formik'
-import TextError from './textError'
+import TextError from './TextError'
 
 function Textarea (props) {
   const { label, name, ...rest } = props
