+++ src/Pages/AppraisalActivity/upload.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Field, ErrorMessage } from 'formik'
-import TextError from './textError'
+import TextError from './TextError'
 
 function Upload (props) {
   const { label, name, ...rest } = props
